% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseCrossVal.R
\name{BaseCrossVal

This an abstract base class providing a common interface for cross-validation.}
\alias{BaseCrossVal

This an abstract base class providing a common interface for cross-validation.}
\title{BaseCrossVal}
\usage{
\dontrun{
  # This is an abstract class, do not instantiate directly.
}
}
\description{
Base class for crossval: kfold.
}
